+++ src/vector.h
@@ -3,14 +3,14 @@
 
 class Vector3{ 
 public:
-    Vector3(double x, double y, double z) : x(x), y(y), z(z) {}
+    Vector3(float x, float y, float z) : x(x), y(y), z(z) {}
     Vector3(const Vector3 &v) : x(v.x), y(v.y), z(v.z) {}
     Vector3() : x(0), y(0), z(0) {}
-    double operator[](int i) const { return (&x)[i]; }
+    float operator[](int i) const { return (&x)[i]; }
 
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 };
 
 int max_component(const Vector3 &v);
@@ -19,7 +19,7 @@ Vector3 operator+(const Vector3 &v1, const Vector3 &v2);
 Vector3 operator-(const Vector3 &v1, const Vector3 &v2);
 Vector3 operator*(const Vector3 &v1, const Vector3 &v2);
 Vector3 operator/(const Vector3 &v1, const Vector3 &v2);
-Vector3 operator/(const Vector3 &v1, double s);
+Vector3 operator/(const Vector3 &v1, float s);
 std::ostream &operator<<(std::ostream &os, const Vector3 &v);
